+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.1] - unreleased
+### Fixed
+- examples/analytic_cuda_device.cu wasn't complete.
+
 ## [0.7.0] - 2023-10-31
 ### Added
 - Expose Rust function `fix_amps_ndarray`